KapoorPaper/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
Pillow>=9.4.0
scikit-learn>=1.2.0
